Choices
Added the choice of how much to split, and also changed how the splitting works. My lovely brain thought it was the cell scaling down and copying itself. It is actually the cell scaling down and copying itself AT the scale of how much it scaled down. Biology and Physics